requirements.txt
streamlit_image_comparison
streamlit_extras
plotly.express
pandas
numpy